+++ include/seqan3/argument_parser/detail/format_man.hpp
@@ -51,8 +51,8 @@ public:
     format_man & operator=(format_man &&) = default;         //!< Defaulted.
     ~format_man() = default;                                 //!< Defaulted.
 
-    //!\copydoc format_help_base(bool)
-    format_man(bool const advanced) : base_type{advanced}
+    //!\copydoc format_help_base(std::vector<std::string> const &, bool const)
+    format_man(std::vector<std::string> const & names, bool const advanced = false) : base_type{names, advanced}
     {};
     //!\}